Загрузка погоды работает

+++ app/src/main/java/com/homework/weatherapp/model/Fact.kt
@@ -6,7 +6,7 @@ import com.google.gson.annotations.SerializedName
 data class Fact(
 
     @SerializedName("cloudness")
-    val cloudness: Int,
+    val cloudness: Double,
     @SerializedName("condition")
     val condition: String,
     @SerializedName("daytime")
@@ -26,7 +26,7 @@ data class Fact(
     @SerializedName("prec_prob")
     val precProb: Int,
     @SerializedName("prec_strength")
-    val precStrength: Int,
+    val precStrength: Double,
     @SerializedName("prec_type")
     val precType: Int,
     @SerializedName("pressure_mm")
